Move config file into molmod directory
